+++ src/twg2/parser/codeParser/extractors/MethodParametersParser.java
@@ -9,6 +9,8 @@ import twg2.ast.interm.annotation.AnnotationSig;
 import twg2.ast.interm.method.ParameterSig;
 import twg2.parser.codeParser.AccessModifier;
 import twg2.parser.codeParser.KeywordUtil;
+import twg2.parser.codeParser.Operator;
+import twg2.parser.codeParser.OperatorUtil;
 import twg2.parser.fragment.CodeToken;
 import twg2.parser.fragment.CodeTokenType;
 import twg2.parser.stateMachine.AstParser;
@@ -21,7 +23,8 @@ import twg2.treeLike.simpleTree.SimpleTree;
 public class MethodParametersParser {
 
 	// TODO does not support default parameters
-	public static List<ParameterSig> extractParamsFromSignature(KeywordUtil<? extends AccessModifier> keywordUtil, AstParser<List<AnnotationSig>> annotationParser, SimpleTree<CodeToken> sigNode) {
+	public static List<ParameterSig> extractParamsFromSignature(KeywordUtil<? extends AccessModifier> keywordUtil, OperatorUtil<? extends Operator> operatorUtil,
+			AstParser<List<AnnotationSig>> annotationParser, SimpleTree<CodeToken> sigNode) {
 		val childs = sigNode.getChildren();
 		int size = childs.size();
 
@@ -64,6 +67,7 @@ public class MethodParametersParser {
 			}
 			i = annotationCompletedAt;
 
+			// read parameter type and name
 			token = childs.get(i + 0).getData();
 
 			val type = token.getText();
@@ -76,10 +80,21 @@ public class MethodParametersParser {
 				// each parameter is expected to have 2, so jump the optional one
 				i++;
 			}
+
+			val name = childs.get(i + 1).getData().getText();
+
+			// read parameter default value if available
+			String defaultValue = null;
+			if(i + 3 < size && (token = childs.get(i + 2).getData()).getTokenType() == CodeTokenType.OPERATOR && operatorUtil.assignmentOperators().is(token)) {
+				if(DataTypeExtractor.isDefaultValueLiteral(token = childs.get(i + 3).getData())) {
+					defaultValue = token.getText();
+					i += 2;
+				}
+			}
+
 			List<AnnotationSig> annotations = annotationParser.getParserResult();
 			annotations = annotations.size() > 0 ? new ArrayList<>(annotations) : null;
-			val name = childs.get(i + 1).getData().getText();
-			val param = new ParameterSig(name, type, paramMods, annotations, optional, null);
+			val param = new ParameterSig(name, type, paramMods, annotations, optional, defaultValue);
 			params.add(param);
 		}