+++ test/twg2/parser/codeParser/test/JavaClassParseTest.java
@@ -4,11 +4,11 @@ import static twg2.parser.test.utils.AnnotationAssert.assertAnnotation;
 import static twg2.parser.test.utils.FieldAssert.assertField;
 import static twg2.parser.test.utils.MethodAssert.assertParameter;
 import static twg2.parser.test.utils.TypeAssert.assertType;
+import static twg2.parser.test.utils.TypeAssert.ary;
 import static twg2.parser.test.utils.TypeAssert.ls;
 
 import java.io.IOException;
 import java.nio.file.Paths;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map.Entry;
@@ -41,7 +41,7 @@ import twg2.treeLike.simpleTree.SimpleTree;
  * @since 2016-1-15
  */
 public class JavaClassParseTest {
-	private static List<String> srcLines = Arrays.asList(
+	private static List<String> srcLines = ls(
 		"package ParserExamples.Samples;",
 		"",
 		"/** A simple class to test parsing.",
@@ -85,7 +85,7 @@ public class JavaClassParseTest {
 		"",
 		"    @WebInvoke(Method = \"PUT\", UriTemplate = \"/SetNames?names={names}\",",
 		"        ResponseFormat = WebMessageFormat.Json)",
-		"    List<Integer> SetNames(final String[] names) {",
+		"    List<Integer> SetNames(Constraints constraints, final String[] names) {",
 		"        new { data = SetNames };",
 		"        : new Dictionary<string, object> { content = value } : new { content = object };",
 		"    }",
@@ -97,7 +97,7 @@ public class JavaClassParseTest {
 	private CodeFileAndAst<JavaBlock> simpleJava = CodeFileAndAst.<JavaBlock>parse(CodeLanguageOptions.JAVA, "SimpleJava.java", "ParserExamples.Samples.SimpleJava", true, srcLines);
 
 	@Parameter
-	private CodeFileSrc file = ParseCodeFile.parseFiles(Arrays.asList(Paths.get("rsc/java/ParserExamples/Models/TrackInfo.java")), FileReadUtil.threadLocalInst(), null).get(0);
+	private CodeFileSrc file = ParseCodeFile.parseFiles(ls(Paths.get("rsc/java/ParserExamples/Models/TrackInfo.java")), FileReadUtil.threadLocalInst(), null).get(0);
 
 
 	public JavaClassParseTest() throws IOException {
@@ -115,7 +115,7 @@ public class JavaClassParseTest {
 		Assert.assertEquals(JavaBlock.CLASS, trackInfoBlock.getBlockType());
 		Assert.assertEquals("TrackInfo", trackInfoBlock.getSignature().getSimpleName());
 
-		Assert.assertEquals(Arrays.asList("Serializable", "Comparable<TrackInfo>"), trackInfoBlock.getSignature().getExtendImplementSimpleNames());
+		Assert.assertEquals(ls("Serializable", "Comparable<TrackInfo>"), trackInfoBlock.getSignature().getExtendImplementSimpleNames());
 	}
 
 
@@ -125,15 +125,15 @@ public class JavaClassParseTest {
 		String fullClassName = simpleJava.fullClassName;
 		Assert.assertEquals(1, blocks.size());
 		ClassAst.SimpleImpl<JavaBlock> clas = blocks.get(0).parsedClass;
-		List<FieldSig> fields = clas.getFields();
-		Assert.assertEquals(5, fields.size());
+		Assert.assertEquals(5, clas.getFields().size());
 
 		Assert.assertEquals(fullClassName, NameUtil.joinFqName(clas.getSignature().getFullName()));
 		Assert.assertEquals(AccessModifierEnum.PUBLIC, clas.getSignature().getAccessModifier());
 		Assert.assertEquals("class", clas.getSignature().getDeclarationType());
 
+		List<FieldSig> fields = clas.getFields();
 		assertField(fields, 0, fullClassName + ".mod", "int");
-		Assert.assertEquals(Arrays.asList(" The modification count. "), fields.get(0).getComments());
+		Assert.assertEquals(ls(" The modification count. "), fields.get(0).getComments());
 		List<AnnotationSig> as = fields.get(0).getAnnotations();
 		// annotations: EmptyAnnotation()
 		assertAnnotation(as, 0, "EmptyAnnotation", new String[0], new String[0]);
@@ -151,7 +151,7 @@ public class JavaClassParseTest {
 		assertAnnotation(as, 6, "MultiNamedArgAnnotation", new String[] { "num", "flag", "value" }, "1.23", "false", "abc");
 
 		assertField(fields, 1, fullClassName + "._name", "String");
-		assertField(fields, 2, fullClassName + ".Names", ls("List", ls("String")));
+		assertField(fields, 2, fullClassName + ".Names", ary("List", ary("String")));
 		assertField(fields, 3, fullClassName + ".Count", "int");
 		assertField(fields, 4, fullClassName + ".accesses", "ZonedDateTime[]");
 
@@ -161,14 +161,14 @@ public class JavaClassParseTest {
 		// AddName(...)
 		MethodSigSimple m = clas.getMethods().get(0);
 		Assert.assertEquals(fullClassName + ".AddName", NameUtil.joinFqName(m.fullName));
-		Assert.assertEquals(Arrays.asList(" Add name\n" +
+		Assert.assertEquals(ls(" Add name\n" +
 		        "     * @param name the name\n" +
 		        "     * @return the names\n" +
 		        "     "), m.comments);
 		List<ParameterSig> ps = m.paramSigs;
 		Assert.assertEquals(2, ps.size());
-		assertParameter(ps, 0, "name", "String", null, Arrays.asList(new AnnotationSig("NotNull", NameUtil.splitFqName("NotNull"), Collections.emptyMap())));
-		assertParameter(ps, 1, "zest", "int", null, null);
+		assertParameter(ps, 0, "name", "String", null, null, ls(new AnnotationSig("NotNull", NameUtil.splitFqName("NotNull"), Collections.emptyMap())));
+		assertParameter(ps, 1, "zest", "int", "0", null, null);
 		// annotations:
 		//{"name": "OperationContract", "arguments": {  } },
 		assertAnnotation(m.annotations, 0, "OperationContract", new String[0], new String[0]);
@@ -178,13 +178,14 @@ public class JavaClassParseTest {
 		assertAnnotation(m.annotations, 2, "TransactionFlow", new String[] { "value" }, new String[] { "TransactionFlowOption.Allowed" });
 
 		//returnType: {"typeName": "Result", "genericParameters": [ {"typeName": "IList", "genericParameters": [ {"typeName": "String"}]}]}
-		assertType(ls("Result", ls("List", ls("String"))), m.returnType);
+		assertType(ary("Result", ary("List", ary("String"))), m.returnType);
 
 		// SetNames(...)
 		m = clas.getMethods().get(1);
 		Assert.assertEquals(fullClassName + ".SetNames", NameUtil.joinFqName(m.fullName));
 		ps = m.paramSigs;
-		assertParameter(ps, 0, "names", "String[]", Arrays.asList(JavaKeyword.FINAL), null);
+		assertParameter(ps, 0, "constraints", "Constraints", null, null, null);
+		assertParameter(ps, 1, "names", "String[]", null, ls(JavaKeyword.FINAL), null);
 	}
 
 }