+++ test/twg2/parser/codeParser/test/JavaModelParseTest.java
@@ -4,10 +4,10 @@ import static twg2.parser.test.utils.AnnotationAssert.assertAnnotation;
 import static twg2.parser.test.utils.FieldAssert.assertField;
 import static twg2.parser.test.utils.MethodAssert.assertParameter;
 import static twg2.parser.test.utils.TypeAssert.assertType;
+import static twg2.parser.test.utils.TypeAssert.ary;
 import static twg2.parser.test.utils.TypeAssert.ls;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 
 import org.junit.Assert;
@@ -32,7 +32,7 @@ import twg2.parser.workflow.CodeFileParsed;
  * @since 2016-1-15
  */
 public class JavaModelParseTest {
-	private static List<String> srcLines = Arrays.asList(
+	private static List<String> srcLines = ls(
 		"package ParserExamples.Samples;",
 		"",
 		"/** A simple class to test parsing.",
@@ -88,14 +88,14 @@ public class JavaModelParseTest {
 		Assert.assertEquals("class", clas.getSignature().getDeclarationType());
 
 		assertField(fields, 0, fullClassName + ".mod", "int");
-		Assert.assertEquals(Arrays.asList(" The modification count. "), fields.get(0).getComments());
+		Assert.assertEquals(ls(" The modification count. "), fields.get(0).getComments());
 		List<AnnotationSig> as = fields.get(0).getAnnotations();
 		// annotations: EmptyAnnotation()
 		assertAnnotation(as, 0, "MultiLineAnnotation", new String[] { "arg1", "arg2", "arg3" }, "alpha-1", "Double.TYPE", "3");
 
 		assertField(fields, 1, fullClassName + "._name", "String");
-		assertField(fields, 2, fullClassName + ".Props", ls("Map", ls("Integer", "String")));
-		assertField(fields, 3, fullClassName + ".hiddenField", ls("List", ls("String")));
+		assertField(fields, 2, fullClassName + ".Props", ary("Map", ary("Integer", "String")));
+		assertField(fields, 3, fullClassName + ".hiddenField", ary("List", ary("String")));
 
 		// methods:
 		Assert.assertEquals(1, clas.getMethods().size());
@@ -103,19 +103,19 @@ public class JavaModelParseTest {
 		// AddName()
 		MethodSigSimple m = clas.getMethods().get(0);
 		Assert.assertEquals(fullClassName + ".SetProps", NameUtil.joinFqName(m.fullName));
-		Assert.assertEquals(Arrays.asList(" Set properties\n" +
+		Assert.assertEquals(ls(" Set properties\n" +
 		        "     * @param props the properties\n" +
 		        "     * @return the properties\n" +
 		        "     "), m.comments);
 		List<ParameterSig> ps = m.paramSigs;
-		assertParameter(ps, 0, "props", "List<String>[]", Arrays.asList(JavaKeyword.FINAL), null);
+		assertParameter(ps, 0, "props", "List<String>[]", null, ls(JavaKeyword.FINAL), null);
 
 		// annotations:
 		// @SetterAnnotation(Prop = "Props", UriTemplate = "/SetProps?props={props}", ResponseFormat = WebMessageFormat.Json)
 		assertAnnotation(m.annotations, 0, "SetterAnnotation", new String[] { "Prop", "UriTemplate", "ResponseFormat" }, new String[] { "Props", "/SetProps?props={props}", "WebMessageFormat.Json" });
 
 		//returnType: {"typeName": "Result", "genericParameters": [ {"typeName": "IList", "genericParameters": [ {"typeName": "String"}]}]}
-		assertType(ls("Result", ls("List", ls("String"))), m.returnType);
+		assertType(ary("Result", ary("List", ary("String"))), m.returnType);
 	}
 
 }